+++ eirc/eirc-registry/src/main/java/ru/flexpay/eirc/registry/entity/ImportErrorType.java
@@ -15,36 +15,36 @@ import java.util.Map;
  */
 public enum ImportErrorType implements IFixedIdType, ILocalizedType {
 
-    CITY_UNRESOLVED(1L),
-    STREET_TYPE_UNRESOLVED(2L),
-    STREET_UNRESOLVED(3L),
-    STREET_AND_BUILDING_UNRESOLVED(4L),
-    BUILDING_UNRESOLVED(5L),
-    APARTMENT_UNRESOLVED(6L),
-    ROOM_UNRESOLVED(19L),
+    CITY_UNRESOLVED(1),
+    STREET_TYPE_UNRESOLVED(2),
+    STREET_UNRESOLVED(3),
+    STREET_AND_BUILDING_UNRESOLVED(4),
+    BUILDING_UNRESOLVED(5),
+    APARTMENT_UNRESOLVED(6),
+    ROOM_UNRESOLVED(19),
 
     /* найдено больше одной записи адреса во внутреннем адресном справочнике */
-    MORE_ONE_CITY(7L),
-    MORE_ONE_STREET_TYPE(8L),
-    MORE_ONE_STREET(9L),
-    MORE_ONE_BUILDING(10L),
-    MORE_ONE_APARTMENT(11L),
-    MORE_ONE_ROOM(20L),
+    MORE_ONE_CITY(7),
+    MORE_ONE_STREET_TYPE(8),
+    MORE_ONE_STREET(9),
+    MORE_ONE_BUILDING(10),
+    MORE_ONE_APARTMENT(11),
+    MORE_ONE_ROOM(20),
 
     /* Найдено более одной записи в коррекциях */
-    MORE_ONE_CITY_CORRECTION(12L),
-    MORE_ONE_STREET_TYPE_CORRECTION(13L),
-    MORE_ONE_STREET_CORRECTION(14L),
-    MORE_ONE_BUILDING_CORRECTION(15L),
-    MORE_ONE_APARTMENT_CORRECTION(16L),
-    MORE_ONE_ROOM_CORRECTION(21L),
+    MORE_ONE_CITY_CORRECTION(12),
+    MORE_ONE_STREET_TYPE_CORRECTION(13),
+    MORE_ONE_STREET_CORRECTION(14),
+    MORE_ONE_BUILDING_CORRECTION(15),
+    MORE_ONE_APARTMENT_CORRECTION(16),
+    MORE_ONE_ROOM_CORRECTION(21),
 
-    ACCOUNT_UNRESOLVED(17L),
-    MORE_ONE_ACCOUNT(18L);
+    ACCOUNT_UNRESOLVED(17),
+    MORE_ONE_ACCOUNT(18);
 
     private static final String RESOURCE_BUNDLE = ImportErrorType.class.getName();
 
-    private Long id;
+    private Integer id;
     private static final Map<LinkStatus, ImportErrorType> mapper = ImmutableMap.<LinkStatus, ImportErrorType>builder().
             put(AddressLinkStatus.CITY_UNRESOLVED, CITY_UNRESOLVED).
             put(AddressLinkStatus.STREET_TYPE_UNRESOLVED, STREET_TYPE_UNRESOLVED).
@@ -67,12 +67,12 @@ public enum ImportErrorType implements IFixedIdType, ILocalizedType {
             put(AddressLinkStatus.MORE_ONE_ROOM_CORRECTION, MORE_ONE_ROOM_CORRECTION).
             build();
 
-    private ImportErrorType(Long id) {
+    private ImportErrorType(Integer id) {
         this.id = id;
     }
 
     @Override
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
